refactor(router): move pre activation helpers into utils
